Update dep lib recastnavigation to TC lastest version (#45)
Update dep lib recastnavigation to TC lastest version

+++ src/server/shared/Memory.h
@@ -21,7 +21,7 @@
 #include "DetourAlloc.h"
 
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
     return (void*)new unsigned char[size];
 }
